feature: implemented verify token handling, to ensure, that user don't see admin panel without auth
