+++ client/src/redux/novaPost/service.js
@@ -1,4 +1,4 @@
-import apiClient from '@config/apiClient';
+import apiClient from '@config/api/apiClient';
 import { createAsyncThunk } from '@reduxjs/toolkit';
 
 const NOVA_POST_SETTLEMENTS_ENDPOINT = import.meta.env.VITE_NOVA_POST_SETTLEMENTS_ENDPOINT;
